requirements.txt
Flask>=2.0.0
Flask-Login>=0.6.0
Flask-WTF>=1.0.0
Werkzeug>=2.0.0
stripe>=5.0.0
reportlab>=3.6.0
python-dotenv>=0.19.0
flask-limiter>=3.8.0
bleach>=6.1.0
Flask-Mail>=0.10.0
sentry-sdk[flask]>=2.0.0
boto3>=1.34.0
pytest-html>=4.1.1
pytest-playwright>=0.5.0
